+++ swim/node.go
@@ -58,7 +58,9 @@ type Options struct {
 func defaultOptions() *Options {
 	opts := &Options{
 		StateTimeouts: StateTimeouts{
-			Suspect: 5000 * time.Millisecond,
+			Suspect:   5 * time.Second,
+			Faulty:    24 * time.Hour,
+			Tombstone: 1 * time.Minute,
 		},
 
 		MinProtocolPeriod: 200 * time.Millisecond,
@@ -187,7 +189,7 @@ func NewNode(app, address string, channel shared.SubChannel, opts *Options) *Nod
 
 	node.memberlist = newMemberlist(node)
 	node.memberiter = newMemberlistIter(node.memberlist)
-	node.stateTransitions = newStateTransitions(node, opts.StateTimeouts)
+	node.stateTransitions = newStateTransitions(node, node.clock, opts.StateTimeouts)
 	node.gossip = newGossip(node, opts.MinProtocolPeriod)
 	node.disseminator = newDisseminator(node)
 	node.rollup = newUpdateRollup(node, opts.RollupFlushInterval,
@@ -379,24 +381,26 @@ func (n *Node) Bootstrap(opts *BootstrapOptions) ([]string, error) {
 //= = = = = = = = = = = = = = = = = = = = = = = = = = = = = = = = = = = = = = =
 
 func (n *Node) handleChanges(changes []Change) {
+	n.disseminator.AdjustMaxPropagations()
 	for _, change := range changes {
 		n.disseminator.RecordChange(change)
 
 		switch change.Status {
 		case Alive:
 			n.stateTransitions.Cancel(change)
-			n.disseminator.AdjustMaxPropagations()
-
-		case Faulty:
-			n.stateTransitions.Cancel(change)
 
 		case Suspect:
 			n.stateTransitions.ScheduleSuspectToFaulty(change)
-			n.disseminator.AdjustMaxPropagations()
+
+		case Faulty:
+			n.stateTransitions.ScheduleFaultyToTombstone(change)
 
 		case Leave:
+			// XXX: should this also reap?
 			n.stateTransitions.Cancel(change)
-			n.disseminator.AdjustMaxPropagations()
+
+		case Tombstone:
+			n.stateTransitions.ScheduleTombstoneToEvict(change)
 		}
 	}
 }